+++ includes/gookto_io/intersection/intersectionPhase.hpp
@@ -7,20 +7,20 @@ class intersectionPhase
 {
 private:
     int id;
-    std::vector<int> connectionRef; 
+    std::vector<long int> connectionRef; 
     std::vector<float> priority;
 
 
 public:
     intersectionPhase(int id);
 
-    void pushConnectionRef(int val);
+    void pushConnectionRef(long int val);
     void pushPriority(float val);
 
     //return connectionRef and Priority
 
-    int getId() { return id; }
-    std::vector<int> getConnRef() { return connectionRef; }
+    long int getId() { return id; }
+    std::vector<long int> getConnRef() { return connectionRef; }
     std::vector<float> getPriority() { return priority; }
